+++ src/index.js
@@ -99,6 +99,7 @@ const components = [
   Input,
   InputNumber,
   Img,
+  ImgCropper,
   Loading,
   Modal,
   Navbar,
@@ -136,8 +137,7 @@ const components = [
   Tooltip,
   Upload,
   MessageBox.wdMessageBox,
-  ImgPreview.wdImgPreview,
-  ImgCropper.wdImgCropper
+  ImgPreview.wdImgPreview
 ]
 
 const install = (Vue, config = {}) => {