+++ examples/docs/docs/changelog.md
@@ -54,6 +54,8 @@
   - 新增 `before-change` 修改前钩子函数  (by [@yawuling](https://github.com/yawuling) )
 - Toast
   - 支持设置 toastKey 来修改 Vue.prototype 原型上的 $toast 属性名 (by [@yawuling](https://github.com/yawuling) )
+- Upload
+  - 新增 `show-preview-list` 控制上传成功后是否展示预览列表  (by [@HXCStudio123](https://github.com/HXCStudio123) )
 
 #### 优化